+++ tests/Extractors/NDJSONTest.php
@@ -1,15 +1,16 @@
 <?php
 
-namespace Rubix\ML\Tests\Datasets\Extractors;
+namespace Rubix\ML\Tests\Extractors;
 
-use Rubix\ML\Datasets\Extractors\NDJSON;
-use Rubix\ML\Datasets\Extractors\Extractor;
+use Rubix\ML\Extractors\NDJSON;
+use Rubix\ML\Extractors\Extractor;
 use PHPUnit\Framework\TestCase;
+use IteratorAggregate;
 
 class NDJSONTest extends TestCase
 {
     /**
-     * @var \Rubix\ML\Datasets\Extractors\NDJSON;
+     * @var \Rubix\ML\Extractors\NDJSON;
      */
     protected $extractor;
 
@@ -22,15 +23,16 @@ class NDJSONTest extends TestCase
     {
         $this->assertInstanceOf(NDJSON::class, $this->extractor);
         $this->assertInstanceOf(Extractor::class, $this->extractor);
+        $this->assertInstanceOf(IteratorAggregate::class, $this->extractor);
     }
 
     public function test_extract() : void
     {
-        $records = $this->extractor->extract();
+        $records = $this->extractor->setOffset(1)->extract();
 
         $expected = [
-            ['nice', 'furry', 'friendly', 4, 'not monster'],
-            ['mean', 'furry', 'loner', -1.5, 'monster'],
+            // ['attitude' => 'nice', 'appearance' => 'furry', 'sociability' => 'friendly', 'rating' => 4, 'class' => 'not monster'],
+            ['attitude' => 'mean', 'appearance' => 'furry', 'sociability' => 'loner', 'rating' => -1.5, 'class' => 'monster'],
             ['nice', 'rough', 'friendly', 2.6, 'not monster'],
             ['mean', 'rough', 'friendly', -1, 'monster'],
             ['nice', 'rough', 'friendly', 2.9, 'not monster'],